+++ tools/api-markdown-documenter/src/renderers/html-renderer/test/Utilities.ts
@@ -8,12 +8,15 @@ import {
 	documentationNodeToHtml,
 	treeFromBody,
 } from "../../../documentation-domain-to-html/index.js";
-import { renderHtml, type RenderDocumentConfig } from "../Render.js";
+import { renderHtml, type RenderDocumentConfiguration } from "../Render.js";
 
 /**
  * Tests the rendering of an individual {@link DocumentationNode}, returning the generated string content.
  */
-export function testRender(node: DocumentationNode, maybeConfig?: RenderDocumentConfig): string {
+export function testRender(
+	node: DocumentationNode,
+	maybeConfig?: RenderDocumentConfiguration,
+): string {
 	const config = maybeConfig ?? {};
 	const html = treeFromBody([documentationNodeToHtml(node, config)], config);
 	return renderHtml(html, config);